requirements.txt
setuptools==49.2.0
opencv-python
click===8.1.3
Flask==2.2.2
itsdangerous>=2.0
MarkupSafe==2.1.1
Werkzeug>=2.2.2
numpy==1.23.2
requests==2.23.0
azure-storage-blob==12.3.1
requests==2.23.0
torch==1.12.0+cpu
torchvision===0.13.0+cpu
matplotlib
pillow
tensorflow==2.9.2
PyYAML>=5.3
scipy==1.9.1
tqdm